refactor: projectNetwork
