+++ src/pages/ContentScripts/index.ts
@@ -5,6 +5,7 @@ import RepoActiInflTrend from '../../anchors/RepoActiInflTrend';
 import PerceptorTab from '../../anchors/PerceptorTab';
 import PerceptorLayout from '../../anchors/PerceptorLayout';
 import DeveloperNetwork from '../../anchors/DeveloperNetwork';
+import ProjectNetwork from '../../anchors/ProjectNetwork';
 import { inject2Perceptor } from './Perceptor';
 initializeIcons();
 inject2Perceptor(DeveloperActiInflTrend);
@@ -13,7 +14,7 @@ inject2Perceptor(RepoActiInflTrend);
 inject2Perceptor(PerceptorTab);
 inject2Perceptor(PerceptorLayout);
 inject2Perceptor(DeveloperNetwork);
-import './ProjectNetwork';
+inject2Perceptor(ProjectNetwork);
 import './Hypertrons';
 import './content.styles.css';
 import { Perceptor } from './Perceptor';
